moved xmtp into notifi-react package
