+++ packages/notifi-react/package.json
@@ -45,7 +45,9 @@
     "date-fns": "^2.29.3",
     "dompurify": "^3.0.5",
     "libphonenumber-js": "^1.10.13",
-    "react-virtuoso": "^3.1.1"
+    "react-virtuoso": "^3.1.1",
+    "@xmtp/consent-proof-signature": "^0.1.2",
+    "@xmtp/react-sdk": "^6.0.1"
   },
   "devDependencies": {
     "@notifi-network/notifi-core": "^0.91.1",
