+++ packages/notifi-react/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "jsx": "react",
-    "lib": ["ES2021", "DOM"]
+    "lib": ["ES2021", "DOM"],
+    "target": "ES2021",
   }
 }
